Compiler changes and extension
I created the list of lines in the compiler, I had to create some functions and things started to get messy so I went ahead and create an extension of the compiler in the helpers folder. I also put the switch dictionary and corresponding functions into the extension, and passed the pline_instance as a parameter to them.